requirements.txt
asttokens==3.0.1
certifi==2026.2.25
charset-normalizer==3.4.7
decorator==5.2.1
executing==2.2.1
glfw==2.10.0
idna==3.11
ipython==9.12.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.2.1
numpy==2.4.4
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
pyglm==2.8.3
Pygments==2.20.0
PyOpenGL==3.1.10
PyOpenGL-accelerate==3.1.10
requests==2.33.1
scipy==1.17.1
stack-data==0.6.3
traitlets==5.14.3
urllib3==2.6.3
wcwidth==0.6.0
